Issue #18669 corrections to align with naming requirements
