+++ dev/io.openliberty.servlet.6.0.internal.factories/src/io/openliberty/servlet/internal/osgi/webapp/factory/package-info.java
@@ -1,0 +1,15 @@
+/*******************************************************************************
+ * Copyright (c) 2022 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+/**
+ * @version 1.0.16
+ */
+@org.osgi.annotation.versioning.Version("1.0.16")
+package io.openliberty.servlet.internal.osgi.webapp.factory;
