+++ dev/io.openliberty.servlet.6.0.internal.factories/src/io/openliberty/servlet/internal/util/factory/URIMatcherFactory60Impl.java
@@ -1,0 +1,32 @@
+/*******************************************************************************
+ * Copyright (c) 2022 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package io.openliberty.servlet.internal.util.factory;
+
+import org.osgi.service.component.annotations.Component;
+
+import com.ibm.wsspi.webcontainer.util.URIMatcher;
+import com.ibm.wsspi.webcontainer.util.URIMatcherFactory;
+import com.ibm.wsspi.webcontainer40.util.URIMatcher40;
+
+@Component(service = URIMatcherFactory.class, property = { "service.vendor=IBM" })
+public class URIMatcherFactory60Impl implements URIMatcherFactory {
+
+    @Override
+    public URIMatcher createURIMatcher() {
+        return new URIMatcher40();
+    }
+
+    @Override
+    public URIMatcher createURIMatcher(boolean scalable) {
+        return new URIMatcher40(scalable);
+    }
+
+}
